refactor: 기존 Interceptor 인증 로직을 Filter로 변환
