+++ src/main/java/nextstep/app/FilterConfig.java
@@ -2,27 +2,40 @@ package nextstep.app;
 
 import nextstep.app.domain.Member;
 import nextstep.app.domain.MemberRepository;
-import nextstep.app.ui.BasicAuthenticationInterceptor;
-import nextstep.app.ui.FormLoginInterceptor;
+import nextstep.security.filter.BasicAuthenticationFilter;
+import nextstep.security.filter.FormLoginAuthenticationFilter;
 import nextstep.security.UserDetails;
 import nextstep.security.UserDetailsService;
+import org.springframework.boot.web.servlet.FilterRegistrationBean;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
 @Configuration
-public class WebConfig implements WebMvcConfigurer {
+public class FilterConfig implements WebMvcConfigurer {
+
     private final MemberRepository memberRepository;
 
-    public WebConfig(MemberRepository memberRepository) {
+    public FilterConfig(MemberRepository memberRepository) {
         this.memberRepository = memberRepository;
     }
 
-    @Override
-    public void addInterceptors(InterceptorRegistry registry) {
-        registry.addInterceptor(new FormLoginInterceptor(userDetailsService())).addPathPatterns("/login");
-        registry.addInterceptor(new BasicAuthenticationInterceptor(userDetailsService())).addPathPatterns("/members");
+    @Bean
+    public FilterRegistrationBean<BasicAuthenticationFilter> basicAuthenticationFilterBean() {
+        FilterRegistrationBean<BasicAuthenticationFilter> registrationBean = new FilterRegistrationBean<>();
+        registrationBean.setFilter(new BasicAuthenticationFilter(userDetailsService()));
+        registrationBean.addUrlPatterns("/members");
+
+        return registrationBean;
+    }
+
+    @Bean
+    public FilterRegistrationBean<FormLoginAuthenticationFilter> formLoginAuthenticationFilterBean() {
+        FilterRegistrationBean<FormLoginAuthenticationFilter> registrationBean = new FilterRegistrationBean<>();
+        registrationBean.setFilter(new FormLoginAuthenticationFilter(userDetailsService()));
+        registrationBean.addUrlPatterns("/login");
+
+        return registrationBean;
     }
 
     @Bean
