+++ src/main/java/nextstep/security/AuthenticationException.java
@@ -1,4 +1,4 @@
-package nextstep.app.ui;
+package nextstep.security;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
